Move some tests from `unit` to `integration`
