+++ tests/unit/test___main__.py
@@ -17,7 +17,6 @@ from ehrql.query_engines.base_sql import BaseSQLQueryEngine
 from ehrql.query_engines.debug import DebugQueryEngine
 from ehrql.query_engines.in_memory import InMemoryQueryEngine
 from ehrql.utils.module_utils import get_sibling_subclasses
-from tests.lib.inspect_utils import function_body_as_string
 
 
 # We just need any old existing file with a ".py" extension for testing purposes, its
@@ -324,48 +323,3 @@ def test_all_backends_have_an_alias():
 def test_all_backend_aliases_match_display_names():
     for alias in BACKEND_ALIASES.keys():
         assert backend_from_id(alias).display_name.lower() == alias
-
-
-def test_debug(capsys, tmp_path):
-    # Verify that the debug subcommand can be invoked.
-    @function_body_as_string
-    def definition():
-        from ehrql import create_dataset
-        from ehrql.tables.core import patients
-
-        dataset = create_dataset()
-        dataset.define_population(patients.date_of_birth.year > 1900)
-
-    definition_path = tmp_path / "show.py"
-    definition_path.write_text(definition)
-
-    dummy_data_path = tmp_path / "dummy-data"
-    dummy_data_path.mkdir()
-    patients_table = dummy_data_path / "patients.csv"
-    patients_table.write_text("patient_id,date_of_birth\n1,2020-10-01")
-    argv = [
-        "debug",
-        str(definition_path),
-        "--dummy-tables",
-        str(dummy_data_path),
-    ]
-    main(argv)
-    captured = capsys.readouterr()
-    assert captured.out == ""
-
-
-def test_debug_rejects_unknown_display_format(capsys, tmp_path):
-    dummy_data_path = tmp_path / "dummy-data"
-    dummy_data_path.mkdir()
-    argv = [
-        "debug",
-        DATASET_DEFINITON_PATH,
-        "--dummy-tables",
-        str(dummy_data_path),
-        "--display-format",
-        "badformat",
-    ]
-    with pytest.raises(SystemExit):
-        main(argv)
-    captured = capsys.readouterr()
-    assert "badformat' is not a supported display format" in captured.err
